requirements.txt
sqlalchemy
alembic
psycopg2-binary
passlib
bcrypt
python-jose
python-multipart 